Make post delete, edit and publish functions work for respective users.

+++ blog/views.py
@@ -37,7 +37,6 @@ def post_new(request):
 		    post.author = request.user
 		    post.save()
 		    return redirect('post_detail', pk=post.pk)
-
 	else:
 		form = PostForm()
 		return render(request, 'blog/post_edit.html', {'form': form})
@@ -49,8 +48,8 @@ def post_edit(request, pk):
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
             post = form.save(commit=False)
-            post.author = request.user
-            post.save()
+            if post.author == request.user:
+                post.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm(instance=post)
@@ -64,13 +63,15 @@ def post_draft_list(request):
 @login_required
 def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.publish()
+    if post.author == request.user:
+        post.publish()
     return redirect('post_list')
 
 @login_required
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
+    if post.author == request.user:
+        post.delete()
     return redirect('post_list')
 
 def add_comment_to_post(request, pk):
